fix(bridge): bounded retry for 5xx and transport errors, extract forward_uid
- Fix crash on 5xx: previously response.raise_for_status() escaped the
  HTTP retry loop (only RequestException was caught inside, SerialException
  outside) and crashed the whole bridge on any 500. Now 5xx and network
  errors retry with exponential backoff up to MAX_HTTP_ATTEMPTS=5
  (~31s window covers brief server restarts) then drop the line so the
  bridge never wedges or crash-loops. 4xx drops immediately (permanent
  rejection).
- Extract forward_uid(api_url, line) for testability; main() now calls it.
- Add ForwardUidTests (5 cases): 2xx accepted, 4xx single-request drop,
  5xx retry-then-drop, 5xx-then-2xx recovery, network error retry.

Co-committed with staged kiosk/deployment work (templates, views,
migrations 0003-0005, static assets, divera_kiosk, systemd units) that
was previously staged but not yet committed.

Tests: 29/29 pass (24 existing + 5 new), manage.py check 0 issues.

+++ rfid_project/attendance/models.py
@@ -3,7 +3,7 @@ from django.utils import timezone
 
 # Create your models here.
 class Student(models.Model):
-    card_id = models.IntegerField()
+    card_id = models.BigIntegerField()
     name = models.CharField(max_length=50, blank=True, null=True)
     dob = models.DateField(blank=True, null=True)
     sex = models.CharField(max_length=7, blank=True, null=True)
@@ -19,7 +19,7 @@ class Student(models.Model):
 
 class Log(models.Model):
     ida = models.IntegerField(default=0)
-    card_id = models.IntegerField()
+    card_id = models.BigIntegerField()
     name = models.CharField(max_length=50)
     date = models.DateField(default=timezone.now)
     time_in = models.TimeField(default=timezone.now)
